[feat/#2] Painter 컴포넌트 분리
